Fix broken make rules resulting in broken build.
Ran into this compiling under Yocto 2.6

* doc/Makefile.in (install): The install target depends on a
prerequisite called html, but html is not the target of any rule.
If it doesn't exist, make will bail with an error message that there is
no rule to make html.  html is a directory that is made as a side effect
of some other rule.  Somehow, this side effect of html/ coming into
existence is not happening in my environment, and so if I remove this
html prerequisite so that the rule actually runs, the cp -r html/*
fails. Putting a defensive if around it.

Resolves: LE-12470
Change-Id: I616a43fb5654e1ab997999ea99c944e1fd8ecade
Signed-off-by: Kaz Kylheku <kkylheku@sierrawireless.com>